requirements.txt
streamlit
pandas
openpyxl
requests
python-dotenv
openai